modules/module-10-workflow-orchestration/requirements.txt
# Python dependencies
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=13.0.0

# Apache Airflow
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.16.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-docker==3.9.2
apache-airflow-providers-http==4.8.0
apache-airflow-providers-ssh==3.9.0

# Database drivers
psycopg2-binary>=2.9.9
pymongo>=4.6.1
redis>=5.0.1

# Data processing
SQLAlchemy>=2.0.25
alembic>=1.13.1

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Data validation
great-expectations>=0.18.8
pandera>=0.17.2

# Workflow utilities
pendulum>=3.0.0
croniter>=2.0.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.3
freezegun>=1.4.0

# Linting & formatting
flake8>=7.0.0
black>=24.1.1
isort>=5.13.2

# Monitoring
statsd>=4.0.1
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.3
requests>=2.31.0
click>=8.1.7

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0

# Jupyter (optional)
jupyter>=1.0.0
jupyterlab>=4.0.10
